# openapi.yaml
openapi: 3.0.0
info:
  title: StudentHub API
  description: Platform for Egyptian students – job matching, video CVs, AI search, discount vouchers
  version: 1.0.0
servers:
  - url: https://tolabco.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, role]
              properties:
                email: { type: string }
                password: { type: string }
                role: { type: string, enum: [student, employer, outlet] }
                full_name: { type: string }
                phone: { type: string }
      responses:
        '201': { description: User registered }
        '400': { description: Missing fields }
        '500': { description: Registration failed }
  /login:
    post:
      summary: Login and get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200': { description: Returns token and user object }
        '401': { description: Invalid credentials }
...
  /apply/{jobId}:
    post:
      security: [{ bearerAuth: [] }]
      summary: Apply to a job (student only)
      parameters:
        - { name: jobId, in: path, required: true, schema: { type: string } }
      responses:
        '201': { description: Application submitted }
        '400': { description: Already applied }
  /my-applications:
    get:
      security: [{ bearerAuth: [] }]
      summary: Student's applications (paginated)
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 10 } }
      responses:
        '200': { description: List of applications }
  /employer/applications:
    get:
      security: [{ bearerAuth: [] }]
      summary: Employer's applicants (paginated)
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 10 } }
      responses:
        '200': { description: List of applicants }
...
  /generate-cv:
    post:
      security: [{ bearerAuth: [] }]
      summary: Generate CV text using DeepSeek AI (student only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [full_name, skills]
              properties:
                full_name: { type: string }
                skills: { type: array, items: { type: string } }
                profession_category: { type: string }
                city: { type: string }
                education: { type: string }
                experience: { type: string }
      responses:
        '200': { description: CV text generated }
  /employer/search-students:
    post:
      security: [{ bearerAuth: [] }]
      summary: Natural language search for students (employer only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query: { type: string }
      responses:
        '200': { description: Filtered student results }
...
  /admin/users:
    get:
      security: [{ bearerAuth: [] }]
      summary: List all users (admin only)
      parameters:
        - { name: page, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer } }
      responses:
        '200': { description: Users list with pagination }
  /admin/users/{userId}/disable:
    put:
      security: [{ bearerAuth: [] }]
      summary: Disable a user (admin only)
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: User disabled }
  /admin/users/{userId}/enable:
    put:
      security: [{ bearerAuth: [] }]
      summary: Enable a user (admin only)
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: User enabled }
  /admin/verification-queue:
    get:
      security: [{ bearerAuth: [] }]
      summary: Get pending verifications (admin only)
      responses:
        '200': { description: List of pending students }
  /admin/verify/{studentId}:
    post:
      security: [{ bearerAuth: [] }]
      summary: Approve student verification (admin only)
      parameters:
        - { name: studentId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Student verified }
  /admin/stats:
    get:
      security: [{ bearerAuth: [] }]
      summary: System statistics (admin only)
      responses:
        '200': { description: Counts of users, jobs, etc. }